cloud_function/streamlit_dashboard/requirements.txt
streamlit==1.35.0
pandas==2.2.2
google-cloud-bigquery==3.24.0
db-dtypes==1.2.0
pyarrow==16.1.0
